Non-zero signatures on other units.
In addition to the Morningstars, several vehicles had a Signature of 1 (meaning they were super-detectable for some reason) and the Beagle had a 2 (making its detection range about 8x normal).  Some variants of some vehicles had a reduced signature, with no indication in the description of why those variants were harder to detect, so I set them to zero too.

The only remaining non-zero Signatures in the mod are the Rattler Fortress, and I kinda don't think that's a problem.