A fix to let the AI use two-turn artillery properly
The AI gets confused about two-turn artillery and either can't fire it at all or stops firing after a single volley, because it's confused about the alternate firing mode for Homing/Copperhead.

This change removes the "Guided" mode setting and changes Homing/Copperhead ammo to bee the same ammoCategory as normal artillery ammo.  Because the normal Artillery ammo has "IsArtillery: True" set on the ammunition itself, it still fires in two-turn mode, and because Homing/Copperhead does not, it still fires same turn.  Homing/Copperhead becomes a selectable ammo type fireable in the normal artillery mode, for players, producing a near-zero difference from the player's perspective, but the AI no longer concludes that the gun is broken/empty and now it keeps firing artillery normally.

In theory you could give the AI Homing/Copperhead ammo and it would be able to use it, but it wouldn't understand the need for TAGs.  It would certainly accept the shots if something happened to BE tagged.